[refactor] CommandBuffer를 CommandBufferData, CommandBufferHandler, CommandBufferOptimizer로 분리
