+++ src/ssd_commands/ssd_command_action.py
@@ -1,5 +1,5 @@
 from src.ssd_file_manager import SSDFileManager
-from src.command_buffer import CommandBuffer
+from src.command_buffer_handler import CommandBufferHandler
 from abc import ABC, abstractmethod
 
 
@@ -10,7 +10,7 @@ class InvalidArgumentException(Exception):
 class SSDCommand(ABC):
     registry = {}
 
-    def __init__(self, ssd_file_manager:SSDFileManager, command_buffer:CommandBuffer, *args):
+    def __init__(self, ssd_file_manager:SSDFileManager, command_buffer:CommandBufferHandler, *args):
         self._ssd_file_manager = ssd_file_manager
         self._command_buffer = command_buffer
         self._arguments = args
